Move authorization.js, session.js to app/session/ dir
