+++ app/index.js
@@ -5,7 +5,7 @@ const morgan = require('morgan');
 const cors = require('cors');
 const bodyParser = require('body-parser');
 const { session } = require('./session');
-const { requireAuthorization } = require('./authorization');
+const { requireAuthorization } = require('./session/index.js');
 const { login } = require('./login');
 const { resourceRequestHandler } = require('./request-data/ob-proxy.js');
 const { OBAccountPaymentServiceProviders } = require('./ob-directory');
